refactor(services): extract web evaluation client
`web-agent-client.ts` kept 218 methods inline while the directory's own
convention — and its Tauri counterpart — composes sibling modules by spread.
Start the split with the evaluation group: `webEvaluationArenas` is touched by
no other method, so the mock state moves with its methods and stays unexported.

The narrow interface follows `BuiltinToolService`: the seven signatures move out
of `AgentService` into `EvaluationService`, which `AgentService` now extends, so
no signature is duplicated and both adapters keep checking against one surface.
`cancelEvaluation` and `exportEvaluation` call `this.getEvaluationArena`, so they
declare `this: AgentService` rather than importing the callee directly — a direct
import would bypass any later override in the composition root.
